components/cspr_summarizers/data_servers/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.2
futures==3.0.5
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.47
typing_extensions==4.5.0

